refactor(oauth): separate HTTP configuration
